# controllers/proto/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "controllers/proto/api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/contacts/auth": {
      "post": {
        "operationId": "Authorize",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protoContactResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoAuthorizeContactRequest"
            }
          }
        ],
        "tags": [
          "ContactService"
        ]
      }
    },
...
  },
  "definitions": {
    "protoAuthorizeContactRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "string"
        },
        "password": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "protoContact": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string",
          "format": "string"
        },
        "email": {
          "type": "string",
          "format": "string"
        },
        "password": {
          "type": "string",
          "format": "string"
        },
        "admin": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "protoContactResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "format": "string"
        },
        "contact": {
          "$ref": "#/definitions/protoContact"
        }
      }
    },
...
  }
}